fix(events): fix event delegation
Event stream filtering for correct matches was previously only being generated
once and caused following events listeners to be generated incorrectly.
